requirements.txt
imageio~=2.31.0
Pillow~=9.5.0
pygame~=2.4.0
